requirements.txt
Django==5.0.6
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
requests
gunicorn
